+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.0] - 2022-08-10
+### Added 
+- Cookie consent "Recently viewed products" to access localStorage legally
+### Changed 
+- Saves by default to sessionStorage
 ## [0.0.4] - 2022-08-10
 ### Changed 
 - Update Bug for advanced pricing > 1000
@@ -19,6 +24,7 @@ All notable changes to this project will be documented in this file.
 ### Initial Release
 
 
+[0.1.0]: https://github.com/BuildIT-Consulting/BuildIT_RecentlyViewedPerformance/compare/0.0.4...0.1.0
 [0.0.4]: https://github.com/BuildIT-Consulting/BuildIT_RecentlyViewedPerformance/compare/0.0.3...0.0.4
 [0.0.3]: https://github.com/BuildIT-Consulting/BuildIT_RecentlyViewedPerformance/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/BuildIT-Consulting/BuildIT_RecentlyViewedPerformance/compare/0.0.1...0.0.2